Fixes rewind issue with requests lib
When a request is redirected by the requests library the data should ideally be rewindable (i.e. have 'tell' and 'seek'). The previous generator implementation did not have that. With this change we enhance the reader we get with the progress callback, but leave any other functions untouched thus fixing the issue.